Bundle gtk locale data
Bundle a patched version of libgtk so that we can pick up
gtk locale data at runtime. This is not possible with
standard builds of libgtk.

Bundle gtk .mo files based on installed .mo files.
Without these files, rtl-locales do not work correctly and
some messages in file picker dialogs are not translated.