Replace upload scripts
Replace the existing upload scripts with
a python wrapper for the release API, and
a more compact shell script calling the wrapper.

The latest artifacts are placed in "continuous",
and also placed under the "continuous-rotating"
release, that contains a rotating history of
recent builds.

Note on semantics: the meaning of "rotate" in the
old upload script referred to existing assets being
replaced by the new ones (if they have the same name).
It now refers to the oldest assets being rotated out
of the list.